cleaned up and added meals controller
