+++ routes/foods.js
@@ -1,6 +1,5 @@
 var express = require("express");
 var router = express.Router();
-const Food = require("../models/food");
 const foodsController = require("../controllers/foodsController");
 
 router.get("/", foodsController.index);
